refactor(web): extract useConfigQuerySync hook
The forward, acl, decap and route pages each carried an identical effect
syncing the selected config into the ?config= URL query param. Hoist it
into a shared useConfigQuerySync hook.
